+++ Examples/OpenGL/JPNGTest/AppDelegate.m
@@ -1,0 +1,23 @@
+//
+//  AppDelegate.m
+//  JPNGTest
+//
+//  Created by Nick Lockwood on 16/01/2013.
+//  Copyright (c) 2013 Charcoal Design. All rights reserved.
+//
+
+#import "AppDelegate.h"
+#import "ViewController.h"
+
+@implementation AppDelegate
+
+- (BOOL)application:(__unused UIApplication *)application didFinishLaunchingWithOptions:(__unused NSDictionary *)launchOptions
+{
+    self.window = [[UIWindow alloc] initWithFrame:[[UIScreen mainScreen] bounds]];
+    self.viewController = [[ViewController alloc] init];
+    self.window.rootViewController = self.viewController;
+    [self.window makeKeyAndVisible];
+    return YES;
+}
+
+@end