+++ Examples/OpenGL/JPNGTest/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  JPNGTest
+//
+//  Created by Nick Lockwood on 16/01/2013.
+//  Copyright (c) 2013 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import <GLKit/GLKit.h>
+
+
+@interface ViewController : GLKViewController
+
+@end